Put OPQ training behind the "train-opq" feature gate
OPQ and Gaussian OPQ training requires LAPACK. This dependency should
not be necessary if one only wants to use a quantizer that was
previously trained. This change puts the training functions of the
OPQs behind a feature gate so that most of the functionality of this
crate can be used without a LAPACK implementation.